Increase networking limits for entities, fix Colosseum trials

+++ HKMP/Game/Client/Entity/Component/ClimberComponent.cs
@@ -14,7 +14,7 @@ internal class ClimberComponent : EntityComponent {
 
     public ClimberComponent(
         NetClient netClient,
-        byte entityId,
+        ushort entityId,
         HostClientPair<GameObject> gameObject,
         Climber climber
     ) : base(netClient, entityId, gameObject) {
