+++ HKMP/Networking/Server/ServerUpdateManager.cs
@@ -280,7 +280,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// <param name="id">The ID of the entity.</param>
     /// <param name="spawningType">The type of the entity that spawned the new entity.</param>
     /// <param name="spawnedType">The type of the entity that was spawned.</param>
-    public void SetEntitySpawn(byte id, EntityType spawningType, EntityType spawnedType) {
+    public void SetEntitySpawn(ushort id, EntityType spawningType, EntityType spawnedType) {
         lock (Lock) {
             PacketDataCollection<EntitySpawn> entitySpawnCollection;
 
@@ -304,7 +304,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <returns>An instance of the entity update in the packet.</returns>
-    private EntityUpdate FindOrCreateEntityUpdate(byte entityId) {
+    private EntityUpdate FindOrCreateEntityUpdate(ushort entityId) {
         EntityUpdate entityUpdate = null;
         PacketDataCollection<EntityUpdate> entityUpdateCollection;
 
@@ -346,7 +346,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="position">The position of the entity.</param>
-    public void UpdateEntityPosition(byte entityId, Vector2 position) {
+    public void UpdateEntityPosition(ushort entityId, Vector2 position) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
 
@@ -360,7 +360,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="scale">The scale data of the entity.</param>
-    public void UpdateEntityScale(byte entityId, EntityUpdate.ScaleData scale) {
+    public void UpdateEntityScale(ushort entityId, EntityUpdate.ScaleData scale) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
 
@@ -375,7 +375,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="animationId">The animation ID of the entity.</param>
     /// <param name="animationWrapMode">The wrap mode of the animation of the entity.</param>
-    public void UpdateEntityAnimation(byte entityId, byte animationId, byte animationWrapMode) {
+    public void UpdateEntityAnimation(ushort entityId, byte animationId, byte animationWrapMode) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
 
@@ -390,7 +390,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="isActive">Whether the entity is active or not.</param>
-    public void UpdateEntityIsActive(byte entityId, bool isActive) {
+    public void UpdateEntityIsActive(ushort entityId, bool isActive) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
 
@@ -404,7 +404,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="data">The list of entity network data to add.</param>
-    public void AddEntityData(byte entityId, List<EntityNetworkData> data) {
+    public void AddEntityData(ushort entityId, List<EntityNetworkData> data) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
 
@@ -419,7 +419,7 @@ internal class ServerUpdateManager : UdpUpdateManager<ClientUpdatePacket, Client
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="fsmIndex">The index of the FSM of the entity.</param>
     /// <param name="data">The host FSM data to add.</param>
-    public void AddEntityHostFsmData(byte entityId, byte fsmIndex, EntityHostFsmData data) {
+    public void AddEntityHostFsmData(ushort entityId, byte fsmIndex, EntityHostFsmData data) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate(entityId);
